Add bias.py module
And renamed:

* SceneParameterSample -> Sampler
* ColorBiasedSceneParameterSampler ->  ColorBiasedSampler
